Extract fetchRecipes into HomePage so that it's refreshed once we navigate to it from IndividualRecipe by pushing it on top of the stack
