+++ components/IndividualRecipe.js
@@ -180,7 +180,7 @@ function IndividualRecipe(props) {
         try {
 
             dispatch(deleteRecipe(recipe.id))
-            props.navigation.navigate('Home')
+            props.navigation.push('Home')
         } catch (error) {
             throw new Error("This is an error");
         }
